Updates licensing library

+++ src/Lithnet.GoogleApps.MA/Enums/Features.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Lithnet.GoogleApps.MA
 {
-    internal enum Features : uint
+    internal enum Features : ulong
     {
         None = 0,
         OrgUnits = 1,
