+++ src/Lithnet.GoogleApps.MA/Enums/Skus.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Lithnet.GoogleApps.MA
 {
-    public enum Skus : uint
+    public enum Skus : ulong
     {
         Standard = 0,
         AddOn = 1,
